requirements.txt
langchain==0.3.20
langchain_openai==0.3.7
langchain_core==0.3.41
streamlit==1.43.0
langchain_community==0.3.19
python-dotenv==1.0.1
pyPDF2==3.0.1
faiss-cpu==1.10.2
streamlit-chat==0.1.1